pyproject.toml
[project]
name = "backend-habibisfinances"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcos Diaz",email = "diazelhalabimarcosa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
